data_collection/requirements.txt
matplotlib
pillow
pathlib
imagesize
numpy
pandas
opencv-python
scikit-learn
selenium
beautifulsoup4
tqdm
